# docs/swagger.yaml
basePath: /book
definitions:
...
  request.OrderBrc20ClaimUpdateReq:
    properties:
      address:
        type: string
      net:
        description: livenet/signet/testnet
        type: string
      orderId:
        type: string
      psbtRaw:
        type: string
    type: object
...
  respond.Brc20ClaimItem:
    properties:
      availableCount:
        description: AvailableCount
        type: integer
      coinAmount:
        description: Brc20 amount
        type: integer
      fee:
        description: claim fee
        type: integer
      inscriptionId:
        description: InscriptionId
        type: string
      net:
        description: Net env
        type: string
      orderId:
        description: Order ID
        type: string
      psbtRaw:
        description: PSBT Raw
        type: string
      tick:
        description: Brc20 symbol
        type: string
    type: object
...
  respond.Message:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      processingTime:
        type: integer
    type: object
...
info:
  contact:
    name: API Support
  description: OrdBook API Service
  title: OrdBook API Service
  version: "1.0"
paths:
...
  /brc20/claim/order:
    get:
      description: Fetch one claim order
      parameters:
      - description: net
        in: query
        name: net
        required: true
        type: string
      - description: tick
        in: query
        name: tick
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Brc20ClaimItem'
      summary: Fetch one claim order
      tags:
      - brc20
  /brc20/claim/order/update:
    post:
      description: Update claim order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderBrc20ClaimUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Update claim order
      tags:
      - brc20
...
schemes:
- https
swagger: "2.0"